requirements.txt
python-dotenv
pillow
openai
numpy
faiss-cpu
torch
transformers
requests
PyMuPDF
python-docx
openpyxl

